+++ packages/flutter/test/material/calendar_date_picker_test.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'feedback_tester.dart';
+import '../widgets/feedback_tester.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
