+++ packages/flutter/test/material/checkbox_list_tile_test.dart
@@ -8,7 +8,8 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'feedback_tester.dart';
+
+import '../widgets/feedback_tester.dart';
 
 Widget wrap({ required Widget child }) {
   return MediaQuery(
