+++ packages/flutter/test/material/filter_chip_test.dart
@@ -13,7 +13,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:leak_tracker_flutter_testing/leak_tracker_flutter_testing.dart';
 
-import 'feedback_tester.dart';
+import '../widgets/feedback_tester.dart';
 
 /// Adds the basic requirements for a Chip.
 Widget wrapForChip({
