+++ packages/flutter/test/material/floating_action_button_test.dart
@@ -13,8 +13,8 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
+import '../widgets/feedback_tester.dart';
 import '../widgets/semantics_tester.dart';
-import 'feedback_tester.dart';
 
 void main() {
   final ThemeData material3Theme = ThemeData(useMaterial3: true);
